requirements.txt
django==3.1
djangorestframework==3.12.4
django-rest-swagger==2.2.0